# digdag-docs/src/_static/swagger.yaml
...
swagger: "2.0"
info:
  description: "Digdag server API"
  version: "0.9.41-SNAPSHOT"
  title: "Digdag"
...
paths:
...
  /api/projects/{id}/schedules:
    get:
      tags:
      - "Project"
      summary: "List schedules of a project with filters"
      description: ""
      operationId: "getSchedules"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "project id"
        required: true
        type: "integer"
        format: "int32"
      - name: "workflow"
        in: "query"
        description: "exact matching filter on workflow name"
        required: false
        type: "string"
      - name: "last_id"
        in: "query"
        description: "list schedules whose id is grater than this id for pagination"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestScheduleCollection"
...
  /api/schedules:
    get:
      tags:
      - "Schedule"
      summary: "List schedules"
      description: ""
      operationId: "getSchedules"
      produces:
      - "application/json"
      parameters:
      - name: "last_id"
        in: "query"
        description: "list schedules whose id is grater than this id for pagination"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestScheduleCollection"
  /api/schedules/{id}:
    get:
      tags:
      - "Schedule"
      summary: "Get a schedule"
      description: ""
      operationId: "getSchedules"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "schedule id"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestSchedule"
  /api/schedules/{id}/backfill:
    post:
      tags:
      - "Schedule"
      summary: "Re-schedule past sessions by count or duration"
      description: ""
      operationId: "backfillSchedule"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "session id"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/RestScheduleBackfillRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestSessionAttemptCollection"
  /api/schedules/{id}/disable:
    post:
      tags:
      - "Schedule"
      summary: "Disable scheduling of new sessions"
      description: ""
      operationId: "disableSchedule"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "session id"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestScheduleSummary"
  /api/schedules/{id}/enable:
    post:
      tags:
      - "Schedule"
      summary: "Re-enable disabled scheduling"
      description: ""
      operationId: "enableSchedule"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "session id"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestScheduleSummary"
  /api/schedules/{id}/skip:
    post:
      tags:
      - "Schedule"
      summary: "Skip future sessions by count or time"
      description: ""
      operationId: "skipSchedule"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "session id"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/RestScheduleSkipRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RestScheduleSummary"
...
definitions:
...
  Config:
    type: "object"
    properties:
      empty:
        type: "boolean"
        default: false
      factory:
        $ref: "#/definitions/ConfigFactory"
      keys:
        type: "array"
        items:
          type: "string"
...
  RestSchedule:
    type: "object"
    properties:
      project:
        $ref: "#/definitions/IdAndName"
      id:
        $ref: "#/definitions/Id"
      disabledAt:
        type: "integer"
        format: "int64"
      nextScheduleTime:
        $ref: "#/definitions/OffsetDateTime"
      nextRunTime:
        type: "integer"
        format: "int64"
      workflow:
        $ref: "#/definitions/IdAndName"
...
  IdAndName:
    type: "object"
    properties:
      name:
        type: "string"
      id:
        $ref: "#/definitions/Id"
  ZoneOffset:
    type: "object"
    properties:
      totalSeconds:
        type: "integer"
        format: "int32"
      id:
        type: "string"
      rules:
        $ref: "#/definitions/ZoneRules"
...
  RestScheduleBackfillRequest:
    type: "object"
    properties:
      fromTime:
        type: "integer"
        format: "int64"
      attemptName:
        type: "string"
      dryRun:
        type: "boolean"
        default: false
      count:
        type: "integer"
        format: "int32"
  ZoneOffsetTransition:
    type: "object"
    properties:
      offsetBefore:
        $ref: "#/definitions/ZoneOffset"
      offsetAfter:
        $ref: "#/definitions/ZoneOffset"
      overlap:
        type: "boolean"
        default: false
      instant:
        type: "integer"
        format: "int64"
      duration:
        $ref: "#/definitions/Duration"
      gap:
        type: "boolean"
        default: false
      dateTimeBefore:
        type: "string"
        format: "date-time"
      dateTimeAfter:
        type: "string"
        format: "date-time"
...
  ZoneOffsetTransitionRule:
    type: "object"
    properties:
      month:
        type: "string"
        enum:
        - "JANUARY"
        - "FEBRUARY"
        - "MARCH"
        - "APRIL"
        - "MAY"
        - "JUNE"
        - "JULY"
        - "AUGUST"
        - "SEPTEMBER"
        - "OCTOBER"
        - "NOVEMBER"
        - "DECEMBER"
      timeDefinition:
        type: "string"
        enum:
        - "UTC"
        - "WALL"
        - "STANDARD"
      standardOffset:
        $ref: "#/definitions/ZoneOffset"
      offsetBefore:
        $ref: "#/definitions/ZoneOffset"
      offsetAfter:
        $ref: "#/definitions/ZoneOffset"
      dayOfMonthIndicator:
        type: "integer"
        format: "int32"
      localTime:
        $ref: "#/definitions/LocalTime"
      midnightEndOfDay:
        type: "boolean"
        default: false
      dayOfWeek:
        type: "string"
        enum:
        - "MONDAY"
        - "TUESDAY"
        - "WEDNESDAY"
        - "THURSDAY"
        - "FRIDAY"
        - "SATURDAY"
        - "SUNDAY"
...
  LocalTime:
    type: "object"
    properties:
      hour:
        type: "integer"
        format: "int32"
      minute:
        type: "integer"
        format: "int32"
      second:
        type: "integer"
        format: "int32"
      nano:
        type: "integer"
        format: "int32"
...
  TemporalUnit:
    type: "object"
    properties:
      duration:
        $ref: "#/definitions/Duration"
      durationEstimated:
        type: "boolean"
        default: false
      dateBased:
        type: "boolean"
        default: false
      timeBased:
        type: "boolean"
        default: false
  RestScheduleSummary:
    type: "object"
    properties:
      id:
        $ref: "#/definitions/Id"
      disabledAt:
        type: "integer"
        format: "int64"
      nextScheduleTime:
        $ref: "#/definitions/OffsetDateTime"
      nextRunTime:
        type: "integer"
        format: "int64"
      createdAt:
        type: "integer"
        format: "int64"
      updatedAt:
        type: "integer"
        format: "int64"
      workflow:
        $ref: "#/definitions/IdAndName"
...
  RestScheduleSkipRequest:
    type: "object"
    properties:
      nextRunTime:
        type: "integer"
        format: "int64"
      fromTime:
        type: "integer"
        format: "int64"
      dryRun:
        type: "boolean"
        default: false
      nextTime:
        $ref: "#/definitions/LocalTimeOrInstant"
      count:
        type: "integer"
        format: "int32"
...
  Duration:
    type: "object"
    properties:
      seconds:
        type: "integer"
        format: "int64"
      nano:
        type: "integer"
        format: "int32"
      zero:
        type: "boolean"
        default: false
      negative:
        type: "boolean"
        default: false
      units:
        type: "array"
        items:
          $ref: "#/definitions/TemporalUnit"
  NameOptionalId:
    type: "object"
    properties:
      name:
        type: "string"
      id:
        $ref: "#/definitions/Id"
  ConfigFactory:
    type: "object"
...
  RestSessionAttempt:
    type: "object"
    properties:
      project:
        $ref: "#/definitions/IdAndName"
      id:
        $ref: "#/definitions/Id"
      createdAt:
        type: "integer"
        format: "int64"
      sessionUuid:
        type: "string"
        format: "uuid"
      retryAttemptName:
        type: "string"
      params:
        $ref: "#/definitions/Config"
      sessionId:
        $ref: "#/definitions/Id"
      finishedAt:
        type: "integer"
        format: "int64"
      sessionTime:
        $ref: "#/definitions/OffsetDateTime"
      workflow:
        $ref: "#/definitions/NameOptionalId"
      success:
        type: "boolean"
        default: false
      done:
        type: "boolean"
        default: false
      cancelRequested:
        type: "boolean"
        default: false
      index:
        type: "integer"
        format: "int32"
  OffsetDateTime:
    type: "object"
    properties:
      offset:
        $ref: "#/definitions/ZoneOffset"
      nano:
        type: "integer"
        format: "int32"
      hour:
        type: "integer"
        format: "int32"
      minute:
        type: "integer"
        format: "int32"
      dayOfYear:
        type: "integer"
        format: "int32"
      dayOfWeek:
        type: "string"
        enum:
        - "MONDAY"
        - "TUESDAY"
        - "WEDNESDAY"
        - "THURSDAY"
        - "FRIDAY"
        - "SATURDAY"
        - "SUNDAY"
      month:
        type: "string"
        enum:
        - "JANUARY"
        - "FEBRUARY"
        - "MARCH"
        - "APRIL"
        - "MAY"
        - "JUNE"
        - "JULY"
        - "AUGUST"
        - "SEPTEMBER"
        - "OCTOBER"
        - "NOVEMBER"
        - "DECEMBER"
      dayOfMonth:
        type: "integer"
        format: "int32"
      year:
        type: "integer"
        format: "int32"
      second:
        type: "integer"
        format: "int32"
      monthValue:
        type: "integer"
        format: "int32"
  LocalTimeOrInstant:
    type: "object"
  RestScheduleCollection:
    type: "object"
    properties:
      schedules:
        type: "array"
        items:
          $ref: "#/definitions/RestSchedule"
...
  ZoneRules:
    type: "object"
    properties:
      fixedOffset:
        type: "boolean"
        default: false
      transitions:
        type: "array"
        items:
          $ref: "#/definitions/ZoneOffsetTransition"
      transitionRules:
        type: "array"
        items:
          $ref: "#/definitions/ZoneOffsetTransitionRule"
  RestSessionAttemptCollection:
    type: "object"
    properties:
      attempts:
        type: "array"
        items:
          $ref: "#/definitions/RestSessionAttempt"
...
  Id:
    type: "object"
...